Bugfix: return correct answer despite multiple riddle requests (#2)
* fix: utilize `callback_id` to reveal correct answer

Problem:
The latest request to the server would store the most-recently fetched `riddle` / `answer` in a global object in an attempt to "remember" the answer to respond to later when the "Reveal Answer" button was clicked. If multiple requests were created using the slash command and one of "Reveal Answer" buttons (not the last button however) was clicked, then the message would be replaced by the most recent `riddle` / `answer`.

Solution:
Instead of randomly selecting a riddle to return on the first request to the slash command, we instead generate a random number (`index`) which is used to select a riddle. `index` is then passed in to the `callback_id`, and when the "Reveal Answer" button is selected we can get the correct  answer by using the `callback_id`.

* Bump version due to bugfix.